+++ lib/kb_ObjectUtilities/kb_ObjectUtilitiesClient.py
@@ -98,6 +98,27 @@ class kb_ObjectUtilities(object):
         return self._client.call_method('kb_ObjectUtilities.KButil_delete_ws_objects',
                                         [params], self._service_ver, context)
 
+    def KButil_undelete_ws_objects(self, params, context=None):
+        """
+        :param params: instance of type "KButil_undelete_ws_objects_Params"
+           (KButil_undelete_ws_objects() ** **  Method for undeleting
+           workspace objects) -> structure: parameter "workspace_name" of
+           type "workspace_name" (** The workspace object refs are of form:
+           ** **    objects = ws.get_objects([{'ref':
+           params['workspace_id']+'/'+params['obj_name']}]) ** ** "ref" means
+           the entire name combining the workspace id and the object name **
+           "id" is a numerical identifier of the workspace or object, and
+           should just be used for workspace ** "name" is a string identifier
+           of a workspace or object.  This is received from Narrative.),
+           parameter "object_types" of list of String, parameter "verbose" of
+           type "bool", parameter "undelete_all" of type "bool"
+        :returns: instance of type "KButil_undelete_ws_objects_Output" ->
+           structure: parameter "report_name" of type "data_obj_name",
+           parameter "report_ref" of type "data_obj_ref"
+        """
+        return self._client.call_method('kb_ObjectUtilities.KButil_undelete_ws_objects',
+                                        [params], self._service_ver, context)
+
     def KButil_update_genome_species_name(self, params, context=None):
         """
         :param params: instance of type