+++ lib/kb_ObjectUtilities/kb_ObjectUtilitiesImpl.py
@@ -49,7 +49,7 @@ class kb_ObjectUtilities:
     ######################################### noqa
     VERSION = "1.2.0"
     GIT_URL = "https://github.com/kbaseapps/kb_ObjectUtilities"
-    GIT_COMMIT_HASH = "207d62740823889865d85ef4fa8e0d6ffd627aff"
+    GIT_COMMIT_HASH = "7cf3e7c2d28735d12b50fa47df531f50364f29f2"
 
     #BEGIN_CLASS_HEADER
     workspaceURL = None
@@ -694,6 +694,106 @@ class kb_ObjectUtilities:
         # At some point might do deeper type checking...
         if not isinstance(returnVal, dict):
             raise ValueError('Method KButil_delete_ws_objects return value ' +
+                             'returnVal is not type dict as required.')
+        # return the results
+        return [returnVal]
+
+    def KButil_undelete_ws_objects(self, ctx, params):
+        """
+        :param params: instance of type "KButil_undelete_ws_objects_Params"
+           (KButil_undelete_ws_objects() ** **  Method for undeleting
+           workspace objects) -> structure: parameter "workspace_name" of
+           type "workspace_name" (** The workspace object refs are of form:
+           ** **    objects = ws.get_objects([{'ref':
+           params['workspace_id']+'/'+params['obj_name']}]) ** ** "ref" means
+           the entire name combining the workspace id and the object name **
+           "id" is a numerical identifier of the workspace or object, and
+           should just be used for workspace ** "name" is a string identifier
+           of a workspace or object.  This is received from Narrative.),
+           parameter "object_types" of list of String, parameter "verbose" of
+           type "bool", parameter "undelete_all" of type "bool"
+        :returns: instance of type "KButil_undelete_ws_objects_Output" ->
+           structure: parameter "report_name" of type "data_obj_name",
+           parameter "report_ref" of type "data_obj_ref"
+        """
+        # ctx is the context object
+        # return variables are: returnVal
+        #BEGIN KButil_undelete_ws_objects
+        console = []
+        invalid_msgs = []
+        updated_object_refs = []
+        self.log(console,'Running KButil_undelete_ws_objects')
+        self.log(console, "\n"+pformat(params))
+        report = ''
+        [OBJID_I, NAME_I, TYPE_I, SAVE_DATE_I, VERSION_I, SAVED_BY_I, WSID_I, WORKSPACE_I, CHSUM_I, SIZE_I, META_I] = range(11)  # object_info tuple
+
+        
+        #### do some basic checks
+        #
+        required_params = ['workspace_name']
+        for req_param in required_params:
+            if not params.get(req_param):
+                raise ValueError('{} parameter is required'.format(req_param))
+
+        # get ws_id
+        ws_id = self.dfuClient.ws_name_to_id(params['workspace_name'])
+                
+        # undelete objects in workspace (OBJID limit is chunk_size*top_iter)
+        total_objs = 0
+        chunk_size = 2000
+        top_iter = 1000
+        for chunk_i in range(0,top_iter):
+            minObjectID = chunk_i * chunk_size
+            if minObjectID == 0:
+                minObjectID = 1
+            maxObjectID = (chunk_i+1) * chunk_size - 1
+            refs_to_undelete = []
+            for obj_id in range(minObjectID, maxObjectID+1):
+                obj_ref = "/".join([str(ws_id),str(obj_id)])
+                refs_to_undelete.append({'ref':obj_ref})
+            try:
+                self.wsClient.undelete_objects(refs_to_undelete)
+            except:
+                for obj_id in range(minObjectID, maxObjectID+1):
+                    obj_ref = "/".join([str(ws_id),str(obj_id)])
+                    try:
+                        self.wsClient.undelete_objects([{'ref':obj_ref}])
+                    except:
+                        break
+                    
+            obj_info_list = self.wsClient.list_objects({
+                'ids':[ws_id],
+                'minObjectID': minObjectID,
+                'maxObjectID': maxObjectID
+            })
+            num_objs = len(obj_info_list)
+            if num_objs < 1:
+                break
+            self.log(console, "chunk {} num objs: {}".format(chunk_i, num_objs))
+            total_objs += num_objs
+
+        msg = "Total objects: {}".format(total_objs)
+        report += msg
+        self.log(console, msg)
+
+        # create report
+        report_info = self.reportClient.create({
+            'workspace_name':params['workspace_name'],
+            'report': {
+                'objects_created':[],
+                'text_message':report
+            }
+        })
+                
+        # Return report and updated_object_refs
+        returnVal = { 'report_name': report_info['name'],
+                      'report_ref': report_info['ref']
+        }
+        #END KButil_undelete_ws_objects
+
+        # At some point might do deeper type checking...
+        if not isinstance(returnVal, dict):
+            raise ValueError('Method KButil_undelete_ws_objects return value ' +
                              'returnVal is not type dict as required.')
         # return the results
         return [returnVal]