+++ test/kb_ObjectUtilities_server_test.py
@@ -380,7 +380,7 @@ class kb_ObjectUtilitiesTest(unittest.TestCase):
     ##
     # HIDE @unittest.skip("skipped test_KButil_delete_ws_objects")
     def test_KButil_delete_ws_objects (self):
-        method = 'KButil_count_ws_objects'
+        method = 'KButil_delete_ws_objects'
         print ("\n\nRUNNING: {}".format(method))
         print ("==================================\n\n")
         [OBJID_I, NAME_I, TYPE_I, SAVE_DATE_I, VERSION_I, SAVED_BY_I, WSID_I, WORKSPACE_I, CHSUM_I, SIZE_I, META_I] = range(11)  # object_info tuple
@@ -407,11 +407,34 @@ class kb_ObjectUtilitiesTest(unittest.TestCase):
             'delete_all': 0
         }
         result = self.getImpl().KButil_delete_ws_objects(self.getContext(),params)[0]
-        #print('RESULT:')
-        #pprint(result)
+        print('RESULT:')
+        pprint(result)
 
         pass
-    
+
+
+    #### test_KButil_undelete_ws_objects():
+    ##
+    # HIDE @unittest.skip("skipped test_KButil_undelete_ws_objects")
+    def test_KButil_undelete_ws_objects (self):
+        method = 'KButil_undelete_ws_objects'
+        print ("\n\nRUNNING: {}".format(method))
+        print ("==================================\n\n")
+        [OBJID_I, NAME_I, TYPE_I, SAVE_DATE_I, VERSION_I, SAVED_BY_I, WSID_I, WORKSPACE_I, CHSUM_I, SIZE_I, META_I] = range(11)  # object_info tuple
+
+        #expected_count = {'KBaseGenomeAnnotations.Assembly': 3,
+        #                  'KBaseGenomes.Genome': 3
+        #                 }
+
+        # run method
+        params = {
+            'workspace_name': self.getWsName()
+        }
+        result = self.getImpl().KButil_undelete_ws_objects(self.getContext(),params)[0]
+        print('RESULT:')
+        pprint(result)
+
+        pass
         
 
     #### test_KButil_update_genome_fields_from_files():